+++ apps/audio_player_basic/firmware/src/config/apb_sam_e70_xult_wm8904_ssc_freertos/usb/usb_msd.h
@@ -156,8 +156,8 @@ typedef struct __attribute__((packed))
         uint8_t value;
         struct __attribute__((packed))
         {
-            unsigned            :7;
-            unsigned direction  :1;
+            uint8_t            :7;
+            uint8_t direction  :1;
         }bits;
     } bmCBWFlags;
